requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0
SQLAlchemy>=2.0.38
Jinja2==3.1.2
click==8.1.7
psycopg2